Add Edit form for articles. Implement handler for article edition

+++ src/main/java/pl/projectarea/project0/user/UserRepository.java
@@ -2,9 +2,11 @@ package pl.projectarea.project0.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import pl.projectarea.project0.user.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     User findByUsername(String username);
+
 }
